backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
psycopg2-binary==2.9.9  # For PostgreSQL support
python-dateutil==2.8.2
gunicorn==21.2.0  # For production deployment